+++ internal/store/memory_store.go
@@ -78,6 +78,6 @@ func (ms *MemoryStore) Close() (err error) {
 	return nil
 }
 
-func (ms *MemoryStore) GetURLsByOwnerID(ctx context.Context, ownerID int64) ([]LinksCouple, error) {
+func (ms *MemoryStore) GetURLsByOwnerID(ctx context.Context, ownerID int) ([]LinksCouple, error) {
 	return nil, errors.New("map store doesn't use this method")
 }
