+++ internal/store/store.go
@@ -24,7 +24,7 @@ type Store interface {
 	// Close закрывает соединение с базой данных
 	Close() (err error)
 	// GetURLsByOwnerID возвращает ссылки по ID пользователя с использованием авторизации.
-	GetURLsByOwnerID(ctx context.Context, ownerID int64) ([]LinksCouple, error)
+	GetURLsByOwnerID(ctx context.Context, ownerID int) ([]LinksCouple, error)
 }
 
 // ErrConflict ошибка, которую используем для сигнала о нарушении целостности данных.
